backend/requirements.txt
APScheduler==3.10.4
alembic==1.16.4
black==25.1.0
fastapi==0.116.1
httpx==0.28.1
numpy==2.3.2
pandas==2.3.1
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
pytest==8.4.1
python-dotenv==1.1.1
ruff==0.12.7
scikit-learn==1.7.1
sqlalchemy==2.0.42
uvicorn[standard]==0.35.0
xgboost==3.0.4

